requirements.txt
flask==2.0.1
pycryptodome==3.10.1
